simplify MLSQL file datasource
