+++ streamingpro-mlsql/src/main/java/streaming/core/datasource/MLSQLBaseFileSource.scala
@@ -1,0 +1,28 @@
+package streaming.core.datasource
+
+import org.apache.spark.sql.{DataFrame, DataFrameReader, DataFrameWriter, Row}
+import streaming.dsl.{DslTool, ScriptSQLExec}
+
+/**
+  * 2019-03-19 WilliamZhu(allwefantasy@gmail.com)
+  */
+abstract class MLSQLBaseFileSource extends MLSQLSource with MLSQLSink with MLSQLSourceInfo with MLSQLRegistry with DslTool {
+  override def load(reader: DataFrameReader, config: DataSourceConfig): DataFrame = {
+    val context = ScriptSQLExec.contextGetOrForTest()
+    val format = config.config.getOrElse("implClass", fullFormat)
+    val owner = config.config.get("owner").getOrElse(context.owner)
+    reader.options(rewriteConfig(config.config)).format(format).load(resourceRealPath(context.execListener, Option(owner), config.path))
+  }
+
+  def rewriteConfig(config: Map[String, String]) = {
+    config
+  }
+
+
+  override def save(writer: DataFrameWriter[Row], config: DataSinkConfig): Unit = {
+    val context = ScriptSQLExec.contextGetOrForTest()
+    val format = config.config.getOrElse("implClass", fullFormat)
+    writer.options(rewriteConfig(config.config)).format(format).save(resourceRealPath(context.execListener, Option(context.owner), config.path))
+  }
+
+}
