+++ streamingpro-mlsql/src/main/java/streaming/core/datasource/impl/MLSQLExcel.scala
@@ -1,33 +1,16 @@
 package streaming.core.datasource.impl
 
 import _root_.streaming.core.datasource._
+import _root_.streaming.dsl.ScriptSQLExec
 import _root_.streaming.dsl.mmlib.algs.param.{BaseParams, WowParams}
-import _root_.streaming.dsl.{DslTool, ScriptSQLExec}
 import org.apache.spark.ml.param.Param
 import org.apache.spark.sql._
 
 /**
   * 2019-02-19 WilliamZhu(allwefantasy@gmail.com)
   */
-class MLSQLExcel(override val uid: String) extends MLSQLSource with MLSQLSink with MLSQLSourceInfo with MLSQLRegistry with WowParams with DslTool {
+class MLSQLExcel(override val uid: String) extends MLSQLBaseFileSource with WowParams {
   def this() = this(BaseParams.randomUID())
-
-  override def load(reader: DataFrameReader, config: DataSourceConfig): DataFrame = {
-    val context = ScriptSQLExec.contextGetOrForTest()
-    val format = config.config.getOrElse("implClass", fullFormat)
-    val owner = config.config.get("owner").getOrElse(context.owner)
-    reader.options(rewriteConfig(config.config)).format(format).load(resourceRealPath(context.execListener, Option(owner), config.path))
-  }
-
-  override def save(writer: DataFrameWriter[Row], config: DataSinkConfig): Unit = {
-    val context = ScriptSQLExec.contextGetOrForTest()
-    val format = config.config.getOrElse("implClass", fullFormat)
-    writer.options(rewriteConfig(config.config)).format(format).save(resourceRealPath(context.execListener, Option(context.owner), config.path))
-  }
-
-  def rewriteConfig(config: Map[String, String]) = {
-    config ++ Map("useHeader" -> config.getOrElse("useHeader", "false"))
-  }
 
   override def sourceInfo(config: DataAuthConfig): SourceInfo = {
     val context = ScriptSQLExec.contextGetOrForTest()
